+++ bocadillo/wsgi.py
@@ -1,5 +1,5 @@
 """WSGI compatibility utilities."""
-from .types import WSGIApp
+from .app_types import WSGIApp
 
 
 def empty_wsgi_app() -> WSGIApp: